Continued tests, and fixed a bug in Vector4d/f.mulAdd: it was forgetting w

+++ test/src/me/anno/tests/geometry/TriangleTest.kt
@@ -30,7 +30,7 @@ class TriangleTest {
     private val numTries = 100
 
     companion object {
-        fun Vector2d.isInsideTriangleBaseline(a: Vector2d, b: Vector2d, c: Vector2d): Boolean {
+        fun Vector2f.isInsideTriangleBaseline(a: Vector2f, b: Vector2f, c: Vector2f): Boolean {
             var sum = 0
             if (getSideSign(a, b) > 0f) sum++
             if (getSideSign(b, c) > 0f) sum++
@@ -39,7 +39,7 @@ class TriangleTest {
             return sum == 0 || sum == 3
         }
 
-        fun Vector2f.isInsideTriangleBaseline(a: Vector2f, b: Vector2f, c: Vector2f): Boolean {
+        fun Vector2d.isInsideTriangleBaseline(a: Vector2d, b: Vector2d, c: Vector2d): Boolean {
             var sum = 0
             if (getSideSign(a, b) > 0f) sum++
             if (getSideSign(b, c) > 0f) sum++
@@ -363,7 +363,8 @@ class TriangleTest {
             val hit = subCrossDot(a, b, c, dir) < 0.0
             val distanceLimit = 1e3f
             val expectedDist = if (hit) distance else Float.POSITIVE_INFINITY
-            val actualDist = rayTriangleIntersectionFront(start, dir, a, b, c, distanceLimit, dstNormal, dstPosition, dstUVW)
+            val actualDist =
+                rayTriangleIntersectionFront(start, dir, a, b, c, distanceLimit, dstNormal, dstPosition, dstUVW)
             assertEquals(min(expectedDist, 1e38f), min(actualDist, 1e38f), 1e-3f)
             if (hit) assertEquals(uvw, dstUVW, 1e-3)
             else assertEquals(dstUVW0, dstUVW) // must not be changed, if the ray doesn't hit
@@ -389,7 +390,8 @@ class TriangleTest {
             val hit = subCrossDot(a, b, c, dir) < 0.0
             val distanceLimit = 1e3
             val expectedDist = if (hit) distance else Double.POSITIVE_INFINITY
-            val actualDist = rayTriangleIntersectionFront(start, dir, a, b, c, distanceLimit, dstNormal, dstPosition, dstUVW)
+            val actualDist =
+                rayTriangleIntersectionFront(start, dir, a, b, c, distanceLimit, dstNormal, dstPosition, dstUVW)
             assertEquals(min(expectedDist, 1e300), min(actualDist, 1e300), 1e-12)
             if (hit) assertEquals(uvw, dstUVW, 1e-11)
             else assertEquals(dstUVW0, dstUVW) // must not be changed, if the ray doesn't hit
